Refactors file structure to use containers directory and legacy directory
