+++ src/components/legacy/billboard.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Video from './video'
+import Video from '../video'
 import { Link } from 'react-router'
 
 const BillBoard = () => (
